Moved the validation methods into their own file and updated code as needed.
