+++ src/app/App.js
@@ -3,100 +3,19 @@ import puzzles from "../api/puzzles.json";
 import { useEffect, useState } from "react";
 import { puzzleConsts as puzzleVals } from "../utils/Constants";
 import { expandBoard, flattenBoard } from "../utils/BoardUtilities";
+import {
+  isValidSquare,
+  isValidRow,
+  isValidColumn,
+  isValidValue,
+} from "../utils/ValidationUtilities";
 
 function App() {
-  const uniqueValues = new Set();
-
   //Flatten puzzle board
   const [puzzleBoard, setPuzzleBoard] = useState([]);
 
   //Flag to know if solver has started
   const [started, setStarted] = useState(false);
-
-  //For a given row and col coordinate, validate the 3x3 square
-  const isValidSquare = (row, col, board) => {
-    uniqueValues.clear();
-
-    //Find the coordinate for the first square in the box
-    let firstRow =
-      Math.floor(row / puzzleVals.squareSize) * puzzleVals.squareSize;
-    let firstColumn =
-      Math.floor(col / puzzleVals.squareSize) * puzzleVals.squareSize;
-
-    //Loop through the items in the square, and validate
-    for (let ri = 0; ri < puzzleVals.squareSize; ri++) {
-      for (let ci = 0; ci < puzzleVals.squareSize; ci++) {
-        //Calculate the index to check
-        let index =
-          (firstRow + ri) * puzzleVals.squareSize * puzzleVals.squareSize +
-          firstColumn +
-          ci;
-        let aItem = board[index];
-
-        //validate the value
-        if (isValidValue(aItem.value)) {
-          if (!uniqueValues.has(aItem.value)) {
-            uniqueValues.add(aItem.value);
-          } else {
-            return false;
-          }
-        }
-      }
-    }
-
-    return true;
-  };
-
-  const isValidRow = (rowIndex, board) => {
-    uniqueValues.clear();
-    if (!board || board.length <= 0 || puzzleVals.maximumValue <= rowIndex)
-      return false;
-
-    let items = board.slice(
-      rowIndex * puzzleVals.maximumValue,
-      rowIndex * puzzleVals.maximumValue + puzzleVals.maximumValue
-    );
-
-    let flag = true;
-    items.forEach((aItem) => {
-      if (isValidValue(aItem.value)) {
-        if (!uniqueValues.has(aItem.value)) {
-          uniqueValues.add(aItem.value);
-        } else {
-          flag = false;
-        }
-      }
-    });
-    return flag;
-  };
-
-  const isValidColumn = (columnIndex, board) => {
-    uniqueValues.clear();
-
-    if (!board || board.length <= 0 || puzzleVals.maximumValue <= columnIndex)
-      return false;
-
-    let items = [];
-    for (let r = 0; r < puzzleVals.maximumValue; r++) {
-      items = [...items, board[r * puzzleVals.maximumValue + columnIndex]];
-    }
-
-    let flag = true;
-    items.forEach((aItem) => {
-      if (isValidValue(aItem.value)) {
-        if (!uniqueValues.has(aItem.value)) {
-          uniqueValues.add(aItem.value);
-        } else {
-          flag = false;
-        }
-      }
-    });
-    return flag;
-  };
-
-  const isValidValue = (value) => {
-    return puzzleVals.minimumValue <= value && value <= puzzleVals.maximumValue;
-  };
 
   const setDisplayValue = (item) => {
     if (item.isOriginal) return item.value;
